refactor: 댓글 영역 리팩토링
