+++ src/features/edit-comment/index.ts
@@ -1,4 +1,5 @@
 export { useCommentEditor, useNewCommentForm, useEditCommentDialog } from "./model/edit-comment.hook";
 export type { NewCommentDraft } from "./model/edit-comment.hook";
-export { CommentAddDialogContainer } from "./ui/CommentAddDialogContainer";
-export { CommentEditDialogContainer } from "./ui/CommentEditDialogContainer";
+export { CommentAddDialogContainer } from "./ui/comment-add-dialog-container";
+export { CommentEditDialogContainer } from "./ui/comment-edit-dialog-container";
+export { CommentsListContainer } from "./ui/comments-list-container";
